Fixed conversion size_t to int
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@1317 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/grid/generation/SurplusVolumeRefinementFunctor.cpp
@@ -6,6 +6,7 @@
 // @author Dirk Pflueger (pflueged@in.tum.de)
 
 #include "grid/generation/SurplusVolumeRefinementFunctor.hpp"
+#include <cmath>
 
 namespace sg
 {
@@ -21,7 +22,7 @@ SurplusVolumeRefinementFunctor::~SurplusVolumeRefinementFunctor()
 
 double SurplusVolumeRefinementFunctor::operator()(GridStorage* storage, size_t seq)
 {
-  return pow(2.0, -storage->get(seq)->getLevelSum())*fabs(alpha->get(seq));
+  return pow(2, - ((int)storage->get(seq)->getLevelSum()))*fabs(alpha->get(seq));
 }
 
 double SurplusVolumeRefinementFunctor::start()
